Add UUID and jsonwebtoken package and move PDF  API endpoints to the correct folder and temporary URL generation
