+++ drizzle/meta/_journal.json
@@ -1,0 +1,13 @@
+{
+  "version": "7",
+  "dialect": "postgresql",
+  "entries": [
+    {
+      "idx": 0,
+      "version": "7",
+      "when": 1749307047241,
+      "tag": "0000_chubby_dragon_man",
+      "breakpoints": true
+    }
+  ]
+}
