+++ example.env
@@ -5,3 +5,7 @@ DATABASE_URL=
 QSTASH_TOKEN=
 QSTASH_CURRENT_SIGNING_KEY=
 QSTASH_NEXT_SIGNING_KEY=
+# URL for QStash webhook
+# Use npx localtunnel --port 3000 to get a public URL locally
+# And get you IP from https://loca.lt/mytunnelpassword
+NEXT_PUBLIC_APP_URL=
